requirements.txt
pandas
djangorestframework-bulk
requests
Django==2.2.3
pytz==2019.1
sqlparse==0.3.0
Django==2.2.3
djangorestframework==3.10.2
pytz==2019.1
sqlparse==0.3.0
